suspension, tyre grip, clutch
(didnt push for a while because i was sick for about 3 weeks) suspension works good enough for now. the tyre grip is also working on the lateral grip. i am now trying to make the longitudinal grip work but i need a engine script first. im now busy with the clutch script